requirements.txt
Pillow
requests
Flask
psycopg2-binary
boto3